Revert "LazerMod 重构"
This reverts commit 030f6540555c2c674328f9b1fd6f082fbfd3554c.

+++ src/main/java/com/now/nowbot/service/messageServiceImpl/TestTypeService.kt
@@ -99,7 +99,7 @@ class TestTypeService(
                 PPMinus3.getInstance(
                         file,
                         if (isChangedRating) {
-                            LazerMod.getModSpeedForStarCalculate(param.modsList).toDouble()
+                            LazerMod.getModSpeedForStarCalculate(param.modsList)
                         } else {
                             param.rate
                         },
